requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
requests==2.31.0
apscheduler==3.10.4
fake-useragent==1.4.0
beautifulsoup4==4.12.2
lxml==4.9.3
retrying==1.3.3
pytz==2024.1
pydantic==2.5.3